+++ src/components/index.js
@@ -1,6 +1,6 @@
 export * from './inits/initDragAndDrop.js'
 export * from './inits/initDeleteCompleted.js'
-// export * from './inits/initAddTodo.js'
+export * from './inits/initAddTodo.js'
 // export * from './inits/initDownload.js'
 // export * from './inits/initChangeStatus.js'
 // export * from './inits/initDelete.js'
